Refactor cli-data
