+++ packages/cli-data/src/modules/commands/list.ts
@@ -1,0 +1,27 @@
+//
+// Copyright 2020 DXOS.org
+//
+
+import { CommandModule } from 'yargs';
+
+import { asyncHandler, print } from '@dxos/cli-core';
+
+import { StateManager } from '../../state-manager';
+
+export const listCommand = (stateManager: StateManager): CommandModule => ({
+  command: ['list'],
+  describe: 'List parties.',
+  builder: yargs => yargs,
+
+  handler: asyncHandler(async (argv: any) => {
+    const { json } = argv;
+
+    const parties = Array.from(stateManager.parties.values()).map(({ partyKey, ...rest }) => ({
+      party: partyKey,
+      ...rest,
+      current: partyKey === stateManager.currentParty
+    }));
+
+    print(parties, { json });
+  })
+});
