+++ packages/cli-data/src/modules/party.test.ts
@@ -1,0 +1,9 @@
+//
+// Copyright 2020 DXOS.org
+//
+
+describe('cli-data: Party', () => {
+  test('Creates a party.', () => {
+
+  });
+});
